fix(advantage): constrain weights to finite non-negative; cast rewards to float
- Config: add `ge=0, allow_inf_nan=False` on both weight fields. Negative
  weights would invert the shaping (reward more tokens); inf/nan poison
  advantages downstream.
- `default_advantage_fn`: cast rewards to float32. The previous
  `dtype=rewards.dtype` for costs truncated fractional weighted sums when
  rollouts emitted integer rewards. Casting also fixes a latent crash on
  int rewards in `_efficiency_shaping`'s `rewards.mean()` (Long unsupported).

+++ src/prime_rl/orchestrator/advantage.py
@@ -53,7 +53,7 @@ def default_advantage_fn(
     letting callers shape on model output, tool-response tokens, or any weighted mix.
     `completion_weight` and `tool_response_weight` are ignored for other `length_penalty` values.
     """
-    rewards = torch.tensor([[r["reward"] for r in group] for group in inputs.rollouts])
+    rewards = torch.tensor([[r["reward"] for r in group] for group in inputs.rollouts], dtype=torch.float32)
 
     if length_penalty == "tokens":
         costs = torch.tensor(
